Separate some test to its own script for testing with ganache
