+++ package.json
@@ -6,8 +6,7 @@
     "compile:contracts": "truffle compile",
     "build": "npm run compile:contracts",
     "pretest": "npm run build",
-    "ganache": "ganache-cli -a 20 -e 1000000",
-    "test-ganache": "truffle test /Users/lisandro/Protofire/KyberNetwork/kyber-pool-master-proxy-contract/test/rewards.js",
+    "test:stress": "npm run pretest && scripts/test.sh",
     "test": "truffle test",
     "lint": "npm run lint:js && npm run lint:sol",
     "lint:js": "eslint .",
